effective test for streamedConverter with process.nextTick

+++ tests/utils/StreamedConverter.test.ts
@@ -1,34 +1,42 @@
 import { StreamedConverter } from '../../src/utils'
 import * as fs from 'node:fs'
 
-const sourcefile: string = 'data_pipeline_tests/data/estat_proj_19np.tsv'
+const sourcefile: string = './data_pipeline_tests/data/estat_proj_19np.tsv'
 const destfile: string = './data_pipeline_tests/data/estat_proj_19np.csv'
 
 describe('Test de la convertion en mode stream', () => {
   beforeEach( () => {
-    if (fs.existsSync(destfile))
-      fs.rmSync(destfile)
+    try {
+      if (fs.existsSync(destfile)) fs.rmSync(destfile)
+    } catch(err) {
+      console.log('error beforeEach: ', err)
+    }
   })
 
   afterAll( () => {
-    if (fs.existsSync(destfile))
-      fs.rmSync(destfile)
+    try {
+      setTimeout(() => fs.rmSync(destfile), 500)
+    } catch(err) {
+      console.log('error afterAll: ', err)
+    }
   })
 
-  it("test de la l'existence du fichier converti avec la methode tofile()", async () => {
+  it("test de l'existence du fichier converti avec la methode toFile()", async () => {
+    
     expect(fs.existsSync(destfile)).toBe(false)
-    const test = new StreamedConverter(
+
+    new StreamedConverter(
       sourcefile, 
       [
         { from: /\\/, to: ','},
         { from: /\t/g, to: ','}
       ],
       false
-    )
+    ).toFile(destfile)
 
-    test.toFile(destfile)
     process.nextTick(
       () => { expect(fs.existsSync(destfile)).toBe(true) }
     )
+    expect.assertions(1)
   })
 })
